chore: delete copy_helper and move code to copy file
